+++ hutool-db/src/main/java/cn/hutool/db/handler/BeanHandler.java
@@ -34,7 +34,6 @@ public class BeanHandler<E> implements RsHandler<E>{
 	public E handle(ResultSet rs) throws SQLException {
 		final ResultSetMetaData  meta = rs.getMetaData();
 		final int columnCount = meta.getColumnCount();
-		
 		return rs.next() ? HandleHelper.handleRow(columnCount, meta, rs, this.elementBeanType) : null;
 	}
 }