+++ hutool-db/src/main/java/cn/hutool/db/handler/HandleHelper.java
@@ -1,6 +1,7 @@
 package cn.hutool.db.handler;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Type;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
@@ -8,9 +9,11 @@ import java.sql.Types;
 import java.util.Collection;
 import java.util.Map;
 
-import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.bean.BeanDesc.PropDesc;
+import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.convert.Convert;
+import cn.hutool.core.lang.Assert;
+import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.ReflectUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.core.util.TypeUtil;
@@ -52,10 +55,30 @@ public class HandleHelper {
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T> T handleRow(int columnCount, ResultSetMetaData meta, ResultSet rs, Class<T> beanClass) throws SQLException {
-		if(Map.class.isAssignableFrom(Entity.class)) {
+		Assert.notNull(beanClass, "Bean Class must be not null !");
+		
+		if(beanClass.isArray()) {
+			//返回数组
+			final Class<?> componentType = beanClass.getComponentType();
+			final Object[] result = ArrayUtil.newArray(componentType, columnCount);
+			for(int i = 0,j = 1; i < columnCount; i++, j++) {
+				result[i] = getColumnValue(rs, j, meta.getColumnType(j), componentType);
+			}
+			return (T) result;
+		} else if(Iterable.class.isAssignableFrom(beanClass)) {
+			//集合
+			final Object[] objRow = handleRow(columnCount, meta, rs, Object[].class);
+			return Convert.convert(beanClass, objRow);
+		} else if(beanClass.isAssignableFrom(Entity.class)) {
+			//Entity的父类都可按照Entity返回
 			return (T) handleRow(columnCount, meta, rs);
+		} else if(String.class == beanClass) {
+			//字符串
+			final Object[] objRow = handleRow(columnCount, meta, rs, Object[].class);
+			return (T) StrUtil.join(", ", objRow);
 		}
 		
+		//普通bean
 		final T bean = ReflectUtil.newInstanceIfPossible(beanClass);
 		//忽略字段大小写
 		final Map<String, PropDesc> propMap = BeanUtil.getBeanDesc(beanClass).getPropMap(true);
@@ -69,13 +92,10 @@ public class HandleHelper {
 			columnLabel = StrUtil.toCamelCase(meta.getColumnLabel(i));
 			type = meta.getColumnType(i);
 			pd = propMap.get(columnLabel);
-			if(null != pd) {
-				setter = pd.getSetter();
-			}
+			setter = (null == pd) ? null : pd.getSetter();
 			if(null != setter) {
-				value = getColumnValue(rs, columnLabel, type);
-				value = Convert.convert(TypeUtil.getFirstParamType(setter), value);
-				ReflectUtil.invoke(bean, setter, new Object[] {value});
+				value = getColumnValue(rs, columnLabel, type, TypeUtil.getFirstParamType(setter));
+				ReflectUtil.invokeWithCheck(bean, setter, new Object[] {value});
 			}
 		}
 		return bean;
@@ -108,7 +128,7 @@ public class HandleHelper {
 	 * @since 3.3.1
 	 */
 	public static <T extends Entity> T handleRow(T row, int columnCount, ResultSetMetaData meta, ResultSet rs, boolean withTableName) throws SQLException {
-		if(withTableName) {
+		if (withTableName) {
 			row.setTableName(meta.getTableName(1));
 		}
 		String columnLabel;
@@ -116,7 +136,7 @@ public class HandleHelper {
 		for (int i = 1; i <= columnCount; i++) {
 			columnLabel = meta.getColumnLabel(i);
 			type = meta.getColumnType(i);
-			row.put(columnLabel, getColumnValue(rs, columnLabel, type));
+			row.put(columnLabel, getColumnValue(rs, columnLabel, type, null));
 		}
 		return row;
 	}
@@ -182,21 +202,65 @@ public class HandleHelper {
 	 * 获取字段值<br>
 	 * 针对日期时间等做单独处理判断
 	 * 
+	 * @param <T> 返回类型
 	 * @param rs {@link ResultSet}
-	 * @param label 字段名
+	 * @param label 字段标签或者字段名
 	 * @param type 字段类型，默认Object
+	 * @param targetColumnType 结果要求的类型，需进行二次转换（null或者Object不转换）
 	 * @return 字段值
 	 * @throws SQLException SQL异常
-	 * @since 3.2.3
 	 */
-	private static Object getColumnValue(ResultSet rs, String label, int type) throws SQLException {
+	private static <T> Object getColumnValue(ResultSet rs, String label, int type, Type targetColumnType) throws SQLException {
+		Object rawValue;
 		switch (type) {
 		case Types.TIMESTAMP:
-			return rs.getTimestamp(label);
+			rawValue = rs.getTimestamp(label);
+			break;
 		case Types.TIME:
-			return rs.getTime(label);
+			rawValue = rs.getTime(label);
+			break;
 		default:
-			return rs.getObject(label);
+			rawValue = rs.getObject(label);
+		}
+		if (null == targetColumnType || Object.class == targetColumnType) {
+			// 无需转换
+			return rawValue;
+		} else {
+			// 按照返回值要求转换
+			return Convert.convert(targetColumnType, rawValue);
+		}
+	}
+
+	/**
+	 * 获取字段值<br>
+	 * 针对日期时间等做单独处理判断
+	 * 
+	 * @param <T> 返回类型
+	 * @param rs {@link ResultSet}
+	 * @param columnIndex 字段索引
+	 * @param type 字段类型，默认Object
+	 * @param targetColumnType 结果要求的类型，需进行二次转换（null或者Object不转换）
+	 * @return 字段值
+	 * @throws SQLException SQL异常
+	 */
+	private static <T> Object getColumnValue(ResultSet rs, int columnIndex, int type, Type targetColumnType) throws SQLException {
+		Object rawValue;
+		switch (type) {
+		case Types.TIMESTAMP:
+			rawValue = rs.getTimestamp(columnIndex);
+			break;
+		case Types.TIME:
+			rawValue = rs.getTime(columnIndex);
+			break;
+		default:
+			rawValue = rs.getObject(columnIndex);
+		}
+		if (null == targetColumnType || Object.class == targetColumnType) {
+			// 无需转换
+			return rawValue;
+		} else {
+			// 按照返回值要求转换
+			return Convert.convert(targetColumnType, rawValue);
 		}
 	}
 	// -------------------------------------------------------------------------------------------------------------- Private method end