+++ hutool-db/src/main/java/cn/hutool/db/handler/StringHandler.java
@@ -5,6 +5,7 @@ import java.sql.SQLException;
 
 /**
  * 处理为字符串结果，当查询结果为单个字符串时使用此处理器
+ * 
  * @author  weibaohui
  */
 public class StringHandler implements RsHandler<String>{