+++ hutool-db/src/test/java/cn/hutool/db/test/EntityTest.java
@@ -4,6 +4,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import cn.hutool.db.Entity;
+import cn.hutool.db.test.pojo.User;
 
 /**
  * Entity测试
@@ -29,25 +30,7 @@ public class EntityTest {
 		Entity entity = Entity.create().set("ID", 2).set("NAME", "testName");
 		User user = entity.toBeanIgnoreCase(User.class);
 		
-		Assert.assertEquals(2, user.getId());
+		Assert.assertEquals(Integer.valueOf(2), user.getId());
 		Assert.assertEquals("testName", user.getName());
-	}
-	
-	public static class User{
-		private int id;
-		private String name;
-		
-		public int getId() {
-			return id;
-		}
-		public void setId(int id) {
-			this.id = id;
-		}
-		public String getName() {
-			return name;
-		}
-		public void setName(String name) {
-			this.name = name;
-		}
 	}
 }