+++ hutool-db/src/test/java/cn/hutool/db/test/pojo/User.java
@@ -1,0 +1,60 @@
+package cn.hutool.db.test.pojo;
+
+/**
+ * 测试用POJO，与测试数据库中的user表对应
+ * 
+ * @author looly
+ *
+ */
+public class User {
+	private Integer id;
+	private String name;
+	private int age;
+	private String birthday;
+	private boolean gender;
+
+	public Integer getId() {
+		return id;
+	}
+
+	public void setId(Integer id) {
+		this.id = id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public int getAge() {
+		return age;
+	}
+
+	public void setAge(int age) {
+		this.age = age;
+	}
+
+	public String getBirthday() {
+		return birthday;
+	}
+
+	public void setBirthday(String birthday) {
+		this.birthday = birthday;
+	}
+
+	public boolean isGender() {
+		return gender;
+	}
+
+	public void setGender(boolean gender) {
+		this.gender = gender;
+	}
+
+	@Override
+	public String toString() {
+		return "User [id=" + id + ", name=" + name + ", age=" + age + ", birthday=" + birthday + ", gender=" + gender + "]";
+	}
+}